docs/COBOL_TRANSPILER_INT/app.py
# Modified app.py from previous COBOL Transpiler implementation
   import re
   import yaml
   from fastapi import FastAPI, UploadFile, Depends, HTTPException
   from fastapi.security import OAuth2PasswordBearer
   from jinja2 import Template
   from pathlib import Path
   from typing import List
   from cryptography.fernet import Fernet
   import logging
   import httpx

   app = FastAPI()
   logging.basicConfig(level=logging.INFO)
   oauth2_scheme = OAuth2PasswordBearer(tokenUrl="auth/token")

   async def authenticate_user(token: str = Depends(oauth2_scheme)):
       if token != "valid-token":  # Replace with LegacyBAAS OAuth2 validation
           raise HTTPException(status_code=403, detail="Unauthorized")
       return {"role": "transpiler"}

   class CobolTranspiler:
       def __init__(self, config_path: str):
           self.config = self._load_config(config_path)
           self.templates = {
               'solidity': Template(Path('templates/solidity.j2').read_text()),
               'corda': Template(Path('templates/corda.kt.j2').read_text()),
               'fabric': Template(Path('templates/fabric.go.j2').read_text())
           }
           self.encryption_key = Fernet.generate_key()
           self.cipher = Fernet(self.encryption_key)

       def _load_config(self, path: str):
           try:
               with open(path, 'r') as f:
                   return yaml.safe_load(f)
           except Exception as e:
               logging.error(f"Config load failed: {e}")
               raise ValueError(f"Invalid config: {str(e)}")

       def parse(self, cobol_content: str):
           try:
               data_division = re.search(
                   r'DATA\s+DIVISION[.\s]*WORKING-STORAGE\s+SECTION(.*?)PROCEDURE\s+DIVISION',
                   cobol_content, re.DOTALL | re.IGNORECASE
               ).group(1)
               procedure_division = re.search(
                   r'PROCEDURE\s+DIVISION[.\s]*(.*?)(?:END PROGRAM|$)', 
                   cobol_content, re.DOTALL | re.IGNORECASE
               ).group(1)
               copybooks = re.findall(r'COPY\s+(\w+)', cobol_content, re.IGNORECASE)
               return {
                   'variables': self._parse_data_division(data_division),
                   'logic': self._parse_procedure(procedure_division),
                   'copybooks': copybooks
               }
           except Exception as e:
               logging.error(f"COBOL parsing failed: {e}")
               raise SyntaxError(f"COBOL parsing failed: {str(e)}")

       def _parse_data_division(self, division: str):
           return [
               {
                   'level': match.group(1),
                   'name': match.group(2).lowerTriad
               .lower(),
                   'type': match.group(3)
               }
               for match in re.finditer(r'(\d+)\s+(\w+)\s+PIC\s+([\w\(\)]+)', division)
           ]

       def _parse_procedure(self, division: str):
           logic = []
           for compute in re.finditer(r'COMPUTE\s+(\w+)\s*=\s*(.*?)\.', division):
               logic.append({
                   'operation': 'compute',
                   'target': compute.group(1).lower(),
                   'expression': compute.group(2)
               })
           for if_stmt in re.finditer(r'IF\s+(.*?)\s+PERFORM\s+(\w+)', division):
               logic.append({
                   'operation': 'if',
                   'condition': if_stmt.group(1),
                   'perform': if_stmt.group(2)
               })
           return logic

       def generate(self, ast, output_path: str):
           core_type = self.config['core_type']
           mapped_vars = self._map_types(ast['variables'])
           code = self.templates[core_type.lower()].render(
               contract_name=Path(output_path).stem,
               variables=mapped_vars,
               logic=ast['logic']
           )
           Path(output_path).write_text(code)
           return code

       def _map_types(self, variables):
           mapped = []
           for var in variables:
               var_type = var['type'].split('(')[0]
               var['output_type'] = self.config['data_types'].get(var_type, 'string')
               if var_type == 'COMP-3':
                   var['output_type'] = 'uint128' if self.config.get('optimize_gas', False) else 'uint256'
               mapped.append(var)
           return mapped

       def generate_tests(self, ast, code: str, core_type: str):
           template = Template(Path('templates/test_solidity.j2').read_text())
           return template.render(
               contract_name=Path(output_path).stem,
               logic=ast['logic']
           )

   @app.post("/banking/transpile")
   async def transpile(
       files: List[UploadFile],
       config: UploadFile,
       user: dict = Depends(authenticate_user)
   ):
       try:
           cobol_content = await files[0].read()
           encrypted_content = CobolTranspiler.cipher.encrypt(cobol_content)
           cobol_content = CobolTranspiler.cipher.decrypt(encrypted_content).decode()
           config_content = yaml.safe_load(await config.read())
           config_path = "temp_config.yaml"
           Path(config_path).write_text(yaml.dump(config_content))

           transpiler = CobolTranspiler(config_path)
           ast = transpiler.parse(cobol_content)
           output_path = f"build/{files[0].filename.split('.')[0]}.sol"
           code = transpiler.generate(ast, output_path)
           tests = transpiler.generate_tests(ast, code, config_content['core_type'])

           # Forward to LegacyBAAS Blockchain Gateway
           async with httpx.AsyncClient() as client:
               response = await client.post(
                   "https://api.legacybaas.com/v1/blockchain/networks/ethereum/transactions",
                   json={
                       "sender": {"address": "0xSenderAddress"},
                       "receiver": {"address": "0xReceiverAddress"},
                       "amount": 0,  # Smart contract deployment
                       "data": code  # Smart contract code
                   },
                   headers={"Authorization": f"Bearer {user['token']}"}
               )
               blockchain_result = response.json()

           logging.info(f"Transpiled {files[0].filename} by user {user['role']}")
           return {
               "code": code,
               "tests": tests,
               "ast": ast,
               "blockchain": blockchain_result
           }
       except Exception as e:
           logging.error(f"Transpilation failed: {e}")
           raise HTTPException(status_code=500, detail=str(e))

   @app.get("/suggest-config")
   async def suggest_config(cobol_snippet: str):
       # Mock Grok integration
       return {
           "core_type": "FIS_IBS",
           "data_types": {"COMP-3": "uint256", "PIC X": "string"},
           "blockchain": "ethereum"
       }