docs/COBOL_TRANSPILER_INT/requirements.txt
fastapi==0.95.0
uvicorn==0.20.0
pyyaml==6.0
jinja2==3.1.2
cryptography==40.0.1
pytest==7.2.2
pytest-asyncio==0.21.0
httpx==0.23.3